requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
mistralai==0.0.8
multidict==6.0.5
openai==1.12.0
orjson==3.9.13
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
python-telegram-bot==20.8
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4